backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.2
SQLAlchemy==2.0.23
bcrypt==4.0.1
groq==0.4.1
vaderSentiment==3.3.2
python-dotenv==1.0.0</content>
<parameter name="filePath">c:\manifesting-motivation-ai\backend\requirements.txt